Changes: Move code to ttq package Fix example code in README
